nxp: lx2160acex7: Add custom handlers for SolidRun Hardware
The current implementations of SR hardware have a GPIO that controls
the power off circuit for the main boards.  We drive this GPIO
on shutdown to hard power off the ATX PSU.

We also implement secure variable storage and set up the SOC so
I2C5 is only accessible in the secure world.

Signed-off-by: Jon Nettleton <jon@solid-run.com>