nxp: lx2160a: Fix hardware prefetching being disabled on core 0
in soc_init() soc_init_percpu() is called before _init_global_data()
however that soc_init_percpu eventually calls _get_global_data, in
order to test if it should disable the cores L2 hardware prefetching.

Since the structure is not initialized yet the behaviour of this check
is indeterminite and was causing a performance issue on core 0 because
the check was returning true.

Signed-off-by: Jon Nettleton <jon@solid-run.com>